+++ libs/mintplayer-ng-bootstrap/file-upload/src/directive/file-upload-template.directive.ts
@@ -10,10 +10,7 @@ export class BsFileUploadTemplateDirective {
 
   constructor() {
     const templateRef = inject(TemplateRef);
-    this.fileUploadComponent.fileTemplate = templateRef;
-
-    // TODO: fileUploadComponent.files is now an input() signal and cannot be assigned from the directive.
-    // Consider converting files to model() on BsFileUploadComponent to restore this functionality.
+    this.fileUploadComponent.fileTemplate.set(templateRef);
   }
 
   readonly bsFileUploadTemplateOf = input<FileUpload[] | undefined>(undefined);